Move RangeInput to global components folder
