+++ client/components/RangeInput.tsx
